requirements.txt
asgiref==3.2.7
bleach==3.1.4
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-classy-tags==1.0.0
django-cors-headers==3.3.0
django-cors-middleware==1.5.0
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-nyt==1.1.5
django-sekizai==1.1.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.3
Pillow==7.0.0
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.1.0
wiki==0.5

oauth2~=1.9.0.post1